+++ src/index.ts
@@ -91,6 +91,13 @@ export namespace Duck {
 
 	export namespace Types {
 		export type GameObject = GameObjectClass;
+		export type Renderable =
+			| GameObjectClass
+			| Duck.Interactive.Button
+			| Duck.Interactive.Text
+			| Duck.Effect
+			| Duck.Lights.StaticLight
+			| Duck.Maps.TileMap;
 		export namespace Game {
 			export interface Config {
 				canvas: HTMLCanvasElement | null;